fix(db): align users.id with sessions FK as bigint unsigned
- Schema: replace serial() with bigint unsigned autoincrement PK on users
- 0000: new installs use bigint for users.id (matches user_id FK)
- 0003: ALTER existing DBs from int to bigint unsigned

Run bun run db:migrate after pull.

+++ drizzle/0000_fine_owl.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `users` (
-	`id` serial AUTO_INCREMENT NOT NULL,
+	`id` bigint unsigned AUTO_INCREMENT NOT NULL,
 	`name` varchar(255) NOT NULL,
 	`email` varchar(255) NOT NULL,
 	`password` varchar(255) NOT NULL,
